collision mur ~ ok - faire animation ?

+++ so_long.h
@@ -6,7 +6,7 @@
 /*   By: user42 <user42@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/08 12:13:20 by malatini          #+#    #+#             */
-/*   Updated: 2021/06/22 13:56:09 by user42           ###   ########.fr       */
+/*   Updated: 2021/06/22 14:21:29 by user42           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -78,8 +78,8 @@ typedef struct	s_map
 typedef struct	s_player
 {
 	t_data	*img;
-	int		x;
-	int		y;
+	float	x;
+	float	y;
 	char	walk_direction;
 	float	move_speed;
 	bool	is_found;
